requirements.txt
python-rtmidi >=1.4.9
keyboard >= 0.13.0